whimpizer-web/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP and CORS
python-multipart==0.0.6
httpx==0.25.2

# Redis for job queue
redis==5.0.1

# Background tasks (alternative to Celery for now)
asyncio

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Existing whimpizer dependencies (inherit from original project)
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
openai==1.3.5
anthropic==0.7.7
google-generativeai==0.3.1
reportlab==4.0.7
Pillow==10.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1